fix(ci): pure utils no longer drag sequelize into frontend vitest — AppError split into middleware/appError.js
The lockstep test loads the backend util chain (publicLuckyDraw →
luckyDraw.js) inside vitest; luckyDraw imported AppError from
errorHandler.js, whose top-line sequelize/pino imports don't resolve in the
frontend CI job (backend deps not installed). AppError now lives in its own
dependency-free module; errorHandler re-exports it so every existing import
keeps working. Verified by running the frontend suite with
backend/node_modules absent (CI simulation): 19 files / 245 green; backend
aiCopyDraft + listingDerivation + luckyDraw suites green through the
re-export.
